Seperate JingleHandler
