+++ Helpers/PlayerStateHandler.cs
@@ -4,7 +4,7 @@ using Archipelago.Core.Util;
 using Archipelago.MultiClient.Net.BounceFeatures.DeathLink;
 using Archipelago.MultiClient.Net.Models;
 using Kokuban;
-using System.Runtime.InteropServices;
+
 
 namespace MedievilArchipelago.Helpers
 {
@@ -44,55 +44,8 @@ namespace MedievilArchipelago.Helpers
 
         public static void KillPlayer()
         {
-            var rnd = new Random();
-            var Sounds = new List<Action>
-            {
-                Beeps,
-                Beeps2,
-                Beeps3
-            };
-            var pick = rnd.Next(Sounds.Count);
-            Sounds[pick]();
-            
+            JingleHandler.DeathJingle();
             Memory.WriteByte(Addresses.GameGlobalScene, 0x06);
-        }
-
-        	 private static void Beeps()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                Console.Beep(1320, 500);
-                Thread.Sleep(125);
-                Console.Beep(1320, 500);
-                Thread.Sleep(125);
-                Console.Beep(1320, 500);
-                Thread.Sleep(125);
-                Console.Beep(2000, 1125);
-            }
-        }
-
-        private static void Beeps2()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                Console.Beep(1350, 300);
-                Thread.Sleep(50);
-                Console.Beep(1050, 300);
-                Thread.Sleep(50);
-                Console.Beep(750, 300);
-                Thread.Sleep(2000);
-            }
-        }
-
-        private static void Beeps3()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                Console.Beep(2000, 250);
-                Thread.Sleep(50);
-                Console.Beep(2000, 250);
-                Thread.Sleep(2450);
-            }
         }
